Weather calculator; prop sites available
In general, this set up updates makes it so that all weather is calculated up front for each method, so that it can be easily queried. This enables future mapping and simplifies the code in some parts. It also enables calculation of the proportion of sites available at each timestep for each method, which I have also included.

Specifically:
ldar_sim - sites are now initialized befoe methods, and each site gets a 'lat index' and 'lon index' variable to tie it to the netcdf4 grid.

ldar_sim_main - minor tweaks

OGI_company - this is where 'deployment days' are calculated during initialization. During each work day, the proportion of sites available is calculated and appended to the timeseries output.

OGI_crew - Greatly simplified weather checking (sites now have index and weather is precalculated - just need to pull the value)

weather_lookup - removed 'get_weather' method as it is no longer necessary. Replaced with 'deployment_days' method which calculates the number of weather-friendly days for a given method, given threshold inputs. Indexing, etc. is removed as this is now done up front for each site.