Check that sites are in bounds
Changes to ldar_sim to check whether each site is within the bounds of the weather data before starting the simulation. This is a simple data input quality check to ensure users haven't made any mistakes. Also updating the AER_sites as some of the coordinates in the previous file were slightly out of bounds by a few decimals!