Fixed bug: Sites missed due to weather higher than total # sites
Previously, if the weather was bad the OGI crews would keep looking at the same sites throughout the day to see if the weather improves (but the weather for a single day is constant). Rather than wasting time, I've added a variable called 'attempted today?' that is set to False at the beginning of each timestep, but can switch to True if the weather is deemed unsuitable. If switched to True, the site won't be attempted again until the following day.