Big update: Costs implemented
For each method, a daily deployment cost can be specified. The cost is only charged if a crew is deployed and visits a minimum of one site.

Also made it so that screening methods count the number of flags they get each day - very informative!

Also fixed a bug in OGI_FU_crew in which every FU crew was visiting every required site - one line of code fixed this so that only one agent/crew visits each flagged site.