include windows packages
Christophe Golkhes site is constantly changing - I've commited the packages that on windows you must install from wheels so we always have access to those specific versions. Except for basemap - that package is too big and has to be downloaded - but it is not necessary for running LDAR sim - only the live demo.