Install basemap from Christophe Golkhe's site
Basemap is too large to commit to the github repository - and pypi has a version from many many years ago - the best option is probably to just get it from Christophe Golkhe's site for now and go from there.

Basemap is only necessary for the live plotter - so if this fails in the future it shouldn't be a huge deal.