Revert mapping
Revert from package-intensive projection mapping back to previous function. Will still need to fix this in the future to either move away from projections (and stay in lat lon), find a better solution, or decide to move ahead with projections and new packages. In the meantime, the code Mo put together for variable epsg codes is preserved in the 'ldar sim public' folder in the file 'generic_functions_projections.py'


Former-commit-id: 4daee6523f41c5c655a8c518a36fd2ae430069bb