Rollover workdays
Simulation much more realistic for facilities that take a long time (even multiple days) to inspect. With this new functionality, crews that do not finish a facility in a day must return to it the next day to finish it... this can go on for as long as it takes (multiple days if necessary).

Only done for OGI and FU_OGI. Doesn't matter as much for screening, and doesn't matter at all for fixed, but should make this functionality generic.