Minimum survey interval change
Minimum survey interval is moved to the site inputs list because it should correspond to survey frequency.  Because different sites can have different survey frequencies, it makes sense that they would also have different minimum intervals. The recommended interval to use is 365/(frequency +1) - 5 for now. In the future, this can become more sophisticated with proper scheduling. The purpose of minimum interval is to ensure that surveys are not too clustered together in time (many regulations require a minimum interval also).


Former-commit-id: 535c2b1b14f328e82212813be1820cffcc9def6e