Softly allow output directory makedirs fail
When running in parallel two processes simultaneously queue a filesystem call to create the output directory and causes a failure.

An alternative solution would be a tiny random sleep so two processes are not running at identical times.

This is a pretty rare bug, but happened this morning - so should be fixed.


Former-commit-id: 43b7ffdda6ca441053e1c01da59695420b171274