fix: Fix E2E test creator handling nested inputs
Reason for change: It was discovered that the git in the test
suite folder was not properly tracking all input files from end-to-end
tests derived from nested input folders not named inputs.

Resolution: All inputs folders are now renamed to inputs when they are
copied over to the test_suite. The global parameters are adjusted
accordingly.

Effects of change: Running end to end test creator with nested input
folders will now result in expected/correct behavior.