feat: Add Site Level Follow-up
Reason for change: Some site level screenings can be
used as a follow up to other site level screenings. This behavior is
currently not implemented in LDAR-Sim.

Resolution: Added logic to visit_site to remove current flags from
sites screened with a follow-up screening and add to them the candidate
flags watch-list again. This results in those sites being flagged for
follow-up again with the component level survey. This change piggybacks
off of the existing follow-up:preferred_method parameter to do this.

Effect of change: No effects on existing code anticipated.

Closes #56 Add site level follow ups