docs: Update LDAr-Sim setup to markdown
Reason for change:
Install instructions required were in plaintext files
which is not inline with other existing LDAR-Sim documentation.
Additionally added instruction on setting up a commit message template.

Resolution:
Updated the setup documentation to be a markdown instead of a text file.
Added instructions on how to configure git to use commit template.

Effect(s) of change: N/A