feat: Add pull request template
Reason for change:
As part of the initiative for cleaner code and process in LDAR-Sim,
we want a way to make pull-requests more uniform and easier to review.

Resolution:
Added a template that will be used when creating pull requests.

Effect(s) of change:
Pull requests will now be based from the Pull Request template