feat: Refactor LDAR-Sim parameter structure. P3.
Reason for change:
This changes is aimed at solving the issue of user confusion when
dealing with LDAR-Sim parameterization, specifically program level
parameters. Many program level parameters are not actually able to be
set at the program level as they are overwritten with the values set
for the baseline program for all other programs. This is misleading as
the parameters being at the program level implies that they can be set
per program.

Resolution:
The majority of program level parameters have been moved to a new
parameter file named virtual_world parameters. There will only exist
one virtual world parameters file per simulation set. This will
provide clarity on the intended purpose of parameters contained within
and enforce the that these parameters cannot be set on a per program
basis for valid LDAR-Sim results to be obtained.

P3 - Refactor of input_manager to read in virtual_world parameters

Effect(s) of change:
BREAKING CHANGE - Previous parameter files will no longer be compatible
with the new version of LDAR-Sim. Functionally,
behavior inside the simulation should be unchanged.