fix: Removed references to subtype_times_file
Reason for change:
With the removal of the subtype times file, references to this removed
functionality must also be removed.

Resolution:
Removed all references in
end-to-end testing suite and the sample subtype_times_file in the
inputs.

Effect(s) of change:
All references to the Subtype times file are now removed.