feat: Remove subtype_times_file
Reason for change:
When using subtype files with pregenerate _leaks, as site are
regenerated, the subtypes times are not carried over.

Resolution:
As site survey times can currently be set in the infrastructure file
on a per site basis AND it is desired to move this functionality over
to the subtype file, the subtype_times_file has been removed and the
functionality will be  reintroduced to the subtype file at a later date.

Effect(s) of change:
subtype_times_file removed