test: Update test suite params to V3
Reason for change:
Tests in the E2E test suite would unable to be run due to using v2
parameters and not passing the parameter version check.

Resolution:
Updated parameters in the E2E test suite to V3 parameterization.

Effect(s) of change:
E2E tests in the test suite can now be run and test results will be
produced.