fix: move test suite dates back to original values
Reason for change:
Test suites were producing different results due to the change in the
default date hypothesized to be causing a butterfly effect shifting the
scheduling of surveys due to leap years. This resulted in all surveys
being moved to a different random seed, and different behavior.

Resolution:
Set the start and end date to the previous default start and end date
in the simulation setting for each test case.

Effect(s) of change:
End-to-end tests are now passing.