docs: Update README to contain acknowledgements
Reason for change:
Previously there was no section of the LDAR-Sim documentation to
contain special thanks or acknowledgements for major contributors. As
LDAR-Sim is open source this is important for transparency of where
LDAR-Sim development funding comes from, but more importantly to
acknowledge the contributions of other groups to LDAR-Sim.

Resolution:
Added a special thanks and acknowledgements section.

Effect(s) of change:
N/A