test: Add unit testing for parameter mapping
Reason for change:
Map parameters and the 2 methods it now calls are functionally
completely new. They need to be tested. IN this case since it's less
of a change to a much larger part of existing LDAR-Sim code and more
of creating new methods, unit testing can be applied.

Resolution:
Added unit tests for map_parameters, handle_parameter_versioning and
map_simulation_settings.

Effect(s) of change:
More unit testing.