perf: Improve weather_lookup_hourly performance
Reason for change:
LDAR-Sim is a complex model and takes time to run simulations.
Improvements to model performance are always desirable.

Resolution:
Changed calls in weather_lookup_hourly class from np.average(array)
to np.mean() array. As np.average() assumes weights of 1 for when
weights are not supplied, this is the same calculation, but avoids
having to do the weight multiplication, which saves time.

Effect(s) of change:
performance improvement