fix: Update sorting watchlist to be saved
Reason for change:
Previously the sorted site watchlist was not being saved, and thus
follow-up scheduling was being done based on recency and not measure
site rates.

While not necessarily wrong from a standpoint of reflecting real world
work practices, this behavior was incorrect in the sense that it did
not match intended behavior or descriptions.

Resolution:
The sorted watchlist is now properly saved to be used for later.

Effect(s) of change:
Simulations using proportion will now function correctly.