feat: Add site summary stats for batch sim running (#116)
* feat: Add site summary stats for batch sim running

Reason for change:
When running simulations in batches, there is a need to summarize and
combine results across multiple simulations.

Resolution:
Some key site statistics are now output to a new site statistics file
when a batch running flag is true/exists. If the flag is true/exists,
columns that wont be useful for combining results are dropped from the
sites csv as well.

Effect(s) of change:
Added batch summary functionality when the batch running flag is true.