feat: Add ts summary stats for batch sim running
Reason for change:
When running simulations in batches, there is a need to summarize and
combine results across multiple simulations.

Resolution:
Some key timeseries statistics are now output to a new timeseries
statistics file when a batch running flag is true/exists.

Effect(s) of change:
Added batch summary functionality for the timeseries when the batch
running flag is true.