fix: add company specific rollover
Reason for change:
This is done to fix an issue where 2 simultaneously deployed mobile
crews from different companies would unintentionally share a rollover,
which would result in erroneous survey/screening times for those
companies and crews.

Resolution:
Added logic to pass a rollover down to crews upon creation by a company.
The rollover is shared by all crews belonging to a company. This means
crews belonging to the same company can coordinate, but crews belonging
to different companies are unable to do so.

Effect(s) of change:
Survey time bug is now resolved