feat: Refactor leaks to emissions
Reason for change:
Leaks need to be refactored to decouple core behaviors of LDAR-sim solve
that new additions to the model functionality do not require sweeping
changes. Additionally, as further emissions logic is planned to be
introduced, leaks need to a subset of total emissions, which should be
supported.

Resolution:
Refactored leaks in LDAR-Sim to use OOP principles, where leaks are
a subclass of emissions.

Effect(s) of change:
BREAKING change!!!