<FEAT>: Sensor module refactor
Reason for change:
This commit addresses the need for an update in the sensor module
following the V4 overhaul of emissions. The existing sensors must
be refactored to accommodate the changes in data structures resulting
from the emission structure rework. Additionally, there is a goal to
rebase the existing code structure to adhere to Object-Oriented
Programming principles.

Resolution:
The changes in the sensor module aim to handle the new data structures
introduced by the emission structure rework. This involves refactoring
existing sensors to align with the updated emissions system.
Furthermore, the codebase is adjusted to follow a more modular and
maintainable structure, adhering to OOP practices.

Effect(s) of change:
- Improved readability and maintainability of the sensor module.
- Inclusion of previous external sensors into the default set
- Improvements to old sensor modules by including default coefficients

This commit lays the groundwork for a more adaptable and extensible
sensor system, accommodating the changes introduced by the emission
structure rework and promoting a more modular codebase.