<FEAT>: V4 Refactor
Reason for change:
The refactoring of different components within LDAR-Sim necessitated
updates to the original run script. As various aspects of the simulation
framework were modified in the V4 refactor, it necessary to bring the
run script in line with these changes.

Resolution:

Effect(s) of change:
- The run script is aligned with the changes made during for the ongoing
 V4 refactor, ensuring compatibility.
- Enhancements from the V4 refactor are now accessible through the
run script, improving overall usability.
- Generator files/pregenerate leaks have been made mandatory
- Pregenerated files are read in as they are used
- State is no longer required for running of LDAR-Sim
- Outputs have not been fully re-implemented
- Multiprocessing is still a WIP