feat: Add emissions and timeseries summary outputs
Reason for change:
LDAR-Sim v4 currently produces timeseries and emissions data outputs
per program per each simulation. With multiple programs and multiple
simulations there currently exists no convenient way to review overall
results. The power of LDAR-SIm lies in running multiple simulations
to approach an "typical" result by considering results
across multiple simulations. This creates a need for results
the individual results of simulations such that the "average"
simulation result can be considered.

Resolution:
Added new outputs reporting:

Program name
Simulation Number
The average Daily Emissions for the simulation
The 95th percentile daily emissions for the simulation
The 5th percentile daily emissions for the simulation
The average Daily Cost for the simulation
The 95th percentile daily Cost for the simulation
The 5th percentile daily Cost for the simulation

for the timeseries output

and

Program name
Simulation Number
Total "True" Emissions for the simulation
The average "True" Emissions for the simulation
The 95th percentile "True" Emissions for the simulation
The 5th percentile "True" Emissions for the simulation
The average emissions rate for the simulation
The 95th percentile emissions rate for the simulation
The 5th percentile emissions rate for the simulation

for the emissions output

Effect(s) of change: New outputs added