<FEAT>: Limit processes
Reason for change:
There is a potential for extra processes to be spawned with  multiprocessing
This resource could be used by the actual processes that are working instead
of being left idle.

Resolution:
If the number of programs is less than the number of available processes, limit
the number of processes that are generated to the number of programs.

Effect(s) of change:
- save memory resource while simulating
- no effect on actual simulation.