feat: Add repairable status to prog emis. summary
Reason for change:
Program emissions summary provides statistics about every emission
generated in simulation. In order to make this as useful as possible to
users, emissions info should include if the emissions was repairable or
not.

Resolution:
Added column indicating if an emission is repairable to the Program
emissions summary.

Effect(s) of change: