<FEAT> Update Equipment Generation
Reason for change:
Original code used a placeholder, integer for equipment generation when
users specify a float or an integer for the number of equipment to generate.

Resolution:
Update the code such that the equipment generation is based on a function
of the emission production rate, to replicate v3 of LDAR-Sim equipment group
functionality.

Effects of change:
- Changed error message to f-string
- Equipment generation is based on emission production rate