<UPD>: Refactor placeholder emission generation
Reason for change:
Placeholder equipment did not take into account emissions production
rates, when determining the number of components to create during
infrastructure initialization.
Having a sufficent number of components such that multiple emissions are
unlikely to occur at the same component which is required emulate version 3 behavior.
In addition, the emissions error handling was done in a non-intuitive
manner, and did not account for all use cases.

Resolution:
Move the error handling to the source class, and refactor the
placeholder equipment creations to use production rates.

Effect(s) of change:
- Change how sources are generated if placeholder equipment are used
- Change how sources handle missing emissions parameters