fix: Improve Emissions Duration Estimates
Reason for change:
Previously, emissions duration estimates were not being calculated
correctly. This change improves the accuracy of these estimates by
accounting for the full duration of th emission, instead of just the
duration until the date of discovery.

Resolution:
This change modifies the emissions duration calculation to account for
the full duration of the emission, instead of just the duration until
the date of discovery.

Effect(s) of change:
Emissions duration estimates will now be more accurate.