<UPD> Update version handling
Reason for change:
The original version handling did not take into account if users were
using parameters that are newer version than the simulator itself.
It always output that the version of parameters are legacy versions
compared to the simulator if the major versions did not match.

Resolution:
Update the version handling to check for newer/older version before
outputting the corresponding message.

Change in code:
- Version handling correct message based on version comparison