<FIX>: Add edge case handling for prog summary
Reason for change:
There was no handling for when there were no fugitive emissions in the
program summary. This caused the simulator to crash when trying to
summarize the simulations.

Resolution:
Add handling for the edge case where there are no fugitive emissions in
simulation

Effect(s) of change:
- Program summary can handle edge case where there are no fugitive
emissions in simulation