feat: Add program mitigation bars visualization
Reason for change:
No visualization was previously outputs conveying program mitigation
amounts. When evaluating LDAR programs it is important to understand the
effectiveness of mitigation efforts.

Resolution:
This change adds mitigation bars to
the summary visualizations to provide a visual representation of the
mitigation efforts of the program.

Effect(s) of change:
Added new output visualization