feat: Add parameters objects to hold parameters
Reason for change:
LDAR-Sim parameters are currently stored in a dictionary.
These parameters are challenging to manage and update, which will be
necessary for the sensitivity analysis module.

Resolution:
Create parameters objects to hold parameters. This will allow for easier
management and updating of parameters.

Effect(s) of change:
N/A