docs: Add Sensitivity Analysis Feature Guide
Reason for change:
The new sensitivity analysis module is powerful, but as a result
setup can be complex and confusing.

Resolution:
Added a new feature guide  that explains how to setup and run the
new sensitivity analysis module.

Effect(s) of change:
Added documentation for the new sensitivity analysis module.