fix: Fixed bug in sens sims variation logic
Reason for change:
A bug was found in the logic for varying parameters to generate
sensitivity analysis simulations. This resulted in some parameter
variations not being properly generated.

Resolution:
Resolved the bug in the logic for varying parameters to generate
sensitivity analysis simulations and added unit tests to ensure
the bug does not reoccur.

Effect(s) of change:
Fixed sensitivity analysis simulation generation logic.