fix: Fix bug missing baseline for sens
Reason for change:
Previously the sensitivity analysis was not able
to run for program or method level because the simulation sets
would be missing a baseline program.
This change ensures the baseline is kept in the simulation set.

Resolution:
The baseline program is now added to the simulation set
during parameter variation.

Effect(s) of change:
Fixed critical bug in sensitivity analysis.