perf: Code runtime optimizations
Reason for change:
Prior to this change, certain scenarios caused runtime of the code
to be significantly longer than expected.

Resolution:
Made runtime optimizations to high use areas of the code to
reduce runtime. A simple test was added to test
gen_estimated_fugitive_emissions_to_remove in order to test the new
code optimizations.

Effect(s) of change:
N/A