<Fix>: Rework program_output
Reason for change:
Program output manager should not be passed into objects that it should
own. Circular references could exist in original code

Resolution:
Move the portion of code that required program manager outside, and
remove the need to pass in the program manager.

Effect(s) of change:
Potential circular referencing is removed.