test: Add simple test for gen_est_fug_to_rem
Reason for change:
There exists a need to test the functionality of
the gen_estimated_fugitive_emissions_to_remove method that is
part of program outputs. This test will ensure
that the function is working as expected.

Resolution:
A simple test was added to test
gen_estimated_fugitive_emissions_to_remove.

Effect(s) of change:
N/A