Add: Sample Emissions Rate Source File
Reason for change:
Emissions rate source file is a required input for the LDAR simulation,
and is complex, which users may find confusing at first.

Resolution:
Added an example emissions rate source file in the inputs folder so
that users can better understand the format of the file.

Effect(s) of change:
N/A