<Fix>: Deployment counts are based on weather
Reason for change:
Previously, even if deployment did not happen due to weather,
it still counted as a deployment. This was not the intended behavior.

Resolution:
Fix the deployment count to only add if weather allowed for deployment

Effect(s) of change:
- Update unit tests to reflect the new behavior
- deployment is now tracked properly and not just incremented if
there was an attempt to send out crew