Refactor: Refactor Emissions Structure
Reason for change:
The Emissions structure did not lend itself to easy use of various
emissions types and internal nomenclature was not consistent with
nomenclature use elsewhere, such as in documentation.

Resolution:
The emissions structure was refactored to be more consistent with
nomenclature used elsewhere in the codebase and to allow for easier
use of various emissions types.

Effect(s) of change:
N/A