feat: Add extrapolation to annual emis calcs
Reason for change:
LDAR-Sim v4 should be able to model uncertainty resulting from
extrapolating emissions from a sample of sites to all sites in
the infrastructure.

Resolution:
This change adds the ability to extrapolate
emissions at sites that are not measured in the sample using
the emissions data from sites with the same site type that are measured.

Effect(s) of change:
Uncertainty in emissions estimates will be improved for
programs measuring only a subset of sites in the total infrastructure.