fix: Resolve Error from having no repairable emis
Reason for change:
An error from attempting to read in empty data was occurring when
simulating a scenario with no repairable emissions.

Resolution:
Added an except block to catch the error and instead
initialize the repairable emissions dataframe with an empty dataframe.

Effect(s) of change:
This change will allow the simulation to run without error when
no repairable emissions are present.