<ENH>: Remove mitigation plots when no mitigation
Reason for change:
Plots that require mitigation values break when there are no mitigation (all emissions are non-repairable in simulation)

Resolution:
Don't plot visualizations that require mitigation numbers when there is no mitigation.

Effect(s) of change:
Added checks for the cost to mitigation plot and the mitigation bar chart to ensure that they are only made when there are values to graph.