<UPD>: Update default parameter data types
Reason for change:
Parameters that should have a float datatype were
set to int datatype. This change updates the
default parameter files to have the correct datatype.

Resolution:
Update the existing parameters to have the correct
datatype

Effect(s) of change:
N/A  - code isn't effected, only the default data
types of the parameters are updated.