fix: Fix error when setting equip properties
Reason for change:
An bug was discovered where users were unable to set properties
in the equipment file due to the way component parsing was handled.

Resolution:
This change fixes the issue by adding more robust parsing of the
equipment file to allow for setting properties.

Effect(s) of change:
This change will allow users to set properties in the equipment file
without encountering an error.