<type>: Short summary of change (limit 50 chars)
Reason for change:
The "stacked cost comparison" visualization is technically a value
visualization, not a cost visualization, as cost is shown as negative.

Resolution:
This change updates the name of the visualization and the y-axis
label to reflect that the visualization is a value visualization.

Effect(s) of change:
Wording now updated to specify the visualization including program
cost is value.
Closes #230