<FIX>: Bugfix to allow for users to set crew count
Reason for change:
A issue was discovered where users could not set crew count in the
settings. This was due to a bug in the code that was not allowing the
user to set the crew count. This is particular affects follow-up methods
as with normal methods, the crew count was determined based on the
survey frequency and average time to complete a survey, whereas
follow-up crew count defaulted to 1.

Resolution:
This change fixes the bug and allows the user to set the crew count.

Effect(s) of change:
Crew counts can now be properly set.