Perf: Performance improvements to detect_emissions
Reason for change:
Profiling indicates that detect_emissions is a high runtime method.

Resolution:
Refactored detect_emissions to reduce runtime, however, a bug
was also discovered and fixed in the process, adding more code
and increasing runtime, resulting in no net change in runtime.

Effect(s) of change:
N/A