fix: Fix edge case for flagging initial sites
Reason for change:
An edge was discovered in which the sites with emissions detected on the
first day of the simulation were not being correctly flagged due to
logic put in place to prevent redundant flagging of sites with tagged
emissions.

Resolution:
The logic was updated to allow for the flagging of sites with emissions
detected on the first day of the simulation.

Effect(s) of change:
This change has a side effect of allowing potentially redundant flagging
of sites in the specific edge case that they receive a tagging and a
flagging survey from 2 different method on the same day. Previously,
this would result in the flagged method being unable to flag the site,
whereas now, it can.