feat: Pass input directory to sensors
Reason for change:
New quantification predictor for sampling quantification error
from a file in the input directory will need the input directory
to be passed through to it in order to read the file.

Resolution:
Added logic to pass the input directory through programs to methods
to sensors, which will then be able to pass it to the quantification
predictor when implemented.

Effect(s) of change:
Sensors now aware of input directory.