<UPD>: Update quantification error initialization
Reason for change:
The original implementation of quantification error did not account for
nan values in the file read. This resulted in a bug where the methods
would eventually detect an emission but have it recorded as nan as
math with nan results in nan values.

Resolution:
Add in handling for nan values in the initialization
Add in other file read error handling

Effect(s) of change:
Update the initialization of quantification error