<ENH>: Remove pregenerated files in script
Reason for change:
Generated pre-seed files can be counter productive when trying to test for changes.
Resolution:
Add into the script to remove the generated infrastructure and emissions files. This forces the generation of new infrastructure and emissions based on the random preseeds, which should result in the same emissions.

Effect(s) of change:
- Pregenerated infrastructure and emissions are deleted with the end to end testing scripts.