test: Update unit tests for optimization change
Reason for change:
Runtime optimization changes to the activate_emissions function in the
generate_emissions module required updates to the unit tests
related to emissions generator as generated emissions are now
expected to be in reverse order.

Resolution:
Inverted the order of the expected emissions in the unit tests.

Effect(s) of change:
All tests now passing.