feat: Refactor end-to-end test case creator script
Reason for change:
The End-to-End test case creator script was long, hard to maintain and
did not share code with other scripts that "run" LDAR-Sim.

Resolution:
Refactored the end-to-end test case creator script to use an end-to-end
test case creator simulation manager that inherits core functionality
from the base simulation manager. This allows the end-to-end test case
creator to inherit changes made to LDAR-Sim run functionality.

Effect(s) of change:
N/A